Matan fix requirements (#9)
* Update .gitignore and correct filename typo

Expanded the .gitignore file to ignore new directories and files in the .idea folder. In addition, corrected a typo in the name of the requirements.txt file.

* Add deployment configuration and update .gitignore

Created new file deployment.xml to include server mapping for different project paths. Updated .gitignore to ignore new files in the .idea folder and GitHub Copilot persisted chat sessions.